Simplify pointer casts.
Function pointers of all types can be converted to one another, and roundtripping results in the original value. So we can write the cast directly, rather than type-punning the containing variable.

(This situation is different from Posix's dlopen, which returns an *object* pointer and requires some amount of type punning.)