[pickups] Allow all pickup types to choose movement type separately.
Previously, we overloaded the item's 'tag' member to describe the
movement type for both "reward" and "goal" objects.

To separate movement from type, pickup items need a new member: When
creating a new pickup, use 'moveType' to describe movement types, and
'typeTag' for the other pickup details.

The member 'tag' aliases to 'moveType' to allow migration. Use of the
old name is deprecated.