[dmlab_so_loader] Call LSAN's end-of-process leak checking routine before dlclose.
This has the effect of running LSAN at that point and then disabling
afterwards. Since we expect the dlclose to happen near the end of the
program life, we still retain many useful leak checks but we avoid
considering unactionable "still-reachable" allocations from within the
DSO to show up as leaks of the main process.