MacOSX support: workspace changes and action changes
We no longer build glib from source, but instead require it to be
available locally.

The new action script uses homebrew for additional library
installations. It only builds a small set of targets that work with
SDL. The //:game_lib_sdl rule is updated to link with `-framework
OpenGL`.

Also update the README.md to state the nature of this branch.