[dmlab_module] Add missing checks for pending errors in the Python runtime.
Any API function that can cause arbitrary callbacks to be called (which
are "step" and "reset") can potentially call back into the Python API
and leave a pending error state.

Therefore, we have to check for this eventuality and propagate the
pending error. Concretely, errors can be set when the level cache
callbacks are themselves calling into the Python runtime.

Similarly, the level cache callbacks themselves need to check for
pending errors, since they may be called arbitrarily from within the
engine.