[maze_generation] Add size constraints to maze generation.
With this change, it is an error to ask for maze generation of a maze size that is too small to contain the largest possible, generated room.

This fixes the observation 'DEBUG.CAMERA_INTERLEAVED.TOP_DOWN', which is now correct for levels `dmlab30/explore_object_rewards_{few,many}`. The observation was previously buggy because the maze width and height (default settings) were too small to contain every possible, generated room. The generated room would be placed regardless of whether it fit, but the resulting entity-layer contained a different formatting. This put the top-down camera in the wrong place.

The affected levels have been updated and will generate layouts similar to before, but the whole maze is offset by 100 units, and object placements will change.