[lua_tests, level_generation] Add further acceptable golden outputs.
This change adds alternative acceptable outputs for test runs of
random algorithms. These tests are inherently poor, because the exact
outcome of the random algorithms was never specified; they are merely
a pragmatic convenience. The actual outputs depend on details of the
standard library implementation, and thus testing with golden outputs
is brittle and prone to break.

However, given that we can't reasonably do much better to test these
algorithms, our pragmatic solution for the time being is to record
outputs from all the implementations we meet and accept any one of
them as a passing test.

Feel free to propose further variants as you encounter them.