[python.BUILD, pip, docs] Extend the Python BUILD rules to allow both Py2 and Py3.
Also update the documentation to match the new Bazel defaults and to
explain the use of Py2 vs Py3 better. This includes an update on how
to build the PIP packaging script.

Finally, the "tested Linux distributions" have been updated to more
recent versions, in order to support Python 3 (and NumPy for Py3). The
previous test results remain valid for Py2 setups, but Py3 is
generally only available on more recent distributions.