[support] Use a more portable implementation of realpath.
This removes the use of a Posix-2008 extension to allow passing NULL
to the output parameter and get a malloc'ed result. According to user
reports, this is not supported on macos 10.15. The alternative is to
rely on PATH_MAX being defined by <limits.h>, which seems to come with
its own share of portability problems, but should at least work on the
platforms we currently support.