[engine] Fix a bug that caused loss of angular velocity information.
The player state update (SetPlayerState) is potentially called
repeatedly for a single frame (e.g. when combined with pixel
observation 'RGB_INTERLEAVED'), and those repeat calls were causing
the angular velocity to be reset (due to the lack of passage of time).
The new behaviour is to ignore repeat calls entirely.