[pip] Update instructions to refer to a map that allows for rewards.
The original instructions that promised a reward of "11.0" were left
over from an early, unreleased state that used a different map that
was never released. The map that we used for the released version,
tests/empty_room_test, does not actually contain any rewards.

This change switches the map in the example over to
demos/extra_entities, which contains four apples in front of the spawn
point. Walking forward for a bit will collect those apples, and thus
produce a reward of 4.0.