[engine] Make tentative definitions of function pointers explicit.
A few OpenGL-related function pointers were accidentally defined
twice, once in sdl_glimp.c and once in (our version of)
tr_extensions.c. This was permitted by common compilers but is an ODR
violation (which is now detected by some other compilers).

The fix is two-fold: first, we replace the superfluous tentative
definitions by declarations. Second, we turn the tentative definitions
that are meant to be canonical into actual definitions by providing
initializers. This will retroactively make the problem detectable even
on the permissive compilers.