[WORKSPACE, tree.BUILD] Use "tree" from head again, custom BUILD
In 5d41ac334cc939c9e1a90158039fce3315f59146 we pinned the dependency
on the "tree" repository to its version 0.1.6, which was the last
version to support Bazel. With this change, we port the last official
Bazel BUILD file to our own, private copy in `bazel/tree.BUILD`.

Our version of the BUILD file had to be adapted slightly, since the
standard http_archive repository rule we use to bring in the
dependency can only place a single BUILD file into the root of the
external repository.

This change is taken directly from
https://github.com/deepmind/lab2d/commit/86637bc89665f0af5f04a1ca8d4b3f522918a71c,
but we also imported pybind11.BUILD from that repository (which is
needed by tree).