[engine] Initialize various PlayerInfo vectors (with zeros).
These values are never meant to be read by code that is within-contract,
but the data is trafficked across the Lua call boundary, and so the values
do end up getting read and written by the Lua binding code. MSAN detects
the uninitialized reads. No user code should be able to observe a difference.